backend/requirements.txt
fastapi
uvicorn
pymongo
requests
typer
python-dotenv